requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
itsdangerous
pytest
redis
email-validator
gunicorn
requests
beautifulsoup4 